hungarian loss layer done

+++ include/caffe/util/hungarian.hpp
@@ -1,27 +1,27 @@
 /********************************************************************
- ********************************************************************
- **
- ** libhungarian by Cyrill Stachniss, 2004
- **
- **
- ** Solving the Minimum Assignment Problem using the
- ** Hungarian Method.
- **
- ** ** This file may be freely copied and distributed! **
- **
- ** Parts of the used code was originally provided by the
- ** "Stanford GraphGase", but I made changes to this code.
- ** As asked by  the copyright node of the "Stanford GraphGase",
- ** I hereby proclaim that this file are *NOT* part of the
- ** "Stanford GraphGase" distrubition!
- **
- ** This file is distributed in the hope that it will be useful,
- ** but WITHOUT ANY WARRANTY; without even the implied
- ** warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
- ** PURPOSE.
- **
- ********************************************************************
- ********************************************************************/
+********************************************************************
+**
+** libhungarian by Cyrill Stachniss, 2004
+**
+**
+** Solving the Minimum Assignment Problem using the
+** Hungarian Method.
+**
+** ** This file may be freely copied and distributed! **
+**
+** Parts of the used code was originally provided by the
+** "Stanford GraphGase", but I made changes to this code.
+** As asked by  the copyright node of the "Stanford GraphGase",
+** I hereby proclaim that this file are *NOT* part of the
+** "Stanford GraphGase" distrubition!
+**
+** This file is distributed in the hope that it will be useful,
+** but WITHOUT ANY WARRANTY; without even the implied
+** warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+** PURPOSE.
+**
+********************************************************************
+********************************************************************/
 
 #ifndef HUNGARIAN_H
 #define HUNGARIAN_H
@@ -40,15 +40,15 @@ extern "C" {
 typedef struct {
   int num_rows;
   int num_cols;
-  int** cost;
+  double** cost;
   int** assignment;
 } hungarian_problem_t;
 
 /** This method initialize the hungarian_problem structure and init
- *  the  cost matrices (missing lines or columns are filled with 0).
- *  It returns the size of the quadratic(!) assignment matrix. **/
+*  the  cost matrices (missing lines or columns are filled with 0).
+*  It returns the size of the quadratic(!) assignment matrix. **/
 int hungarian_init(hungarian_problem_t* p,
-                   int** cost_matrix,
+                   double** cost_matrix,
                    int rows,
                    int cols,
                    int mode);
@@ -68,10 +68,11 @@ void hungarian_print_costmatrix(hungarian_problem_t* p);
 /** Print cost matrix and assignment matrix. **/
 void hungarian_print_status(hungarian_problem_t* p);
 
-int** array_to_matrix(int* m, int rows, int cols);
+double** array_to_matrix(double* m, int rows, int cols);
 
 #ifdef __cplusplus
 }
 #endif
 
 #endif
+
